fix: migrate from using size_t to uint32_t when calculating packet size

+++ src/msgs/sensor_msgs/msg/laser_scan.cpp
@@ -10,10 +10,10 @@ namespace tcp_ip_bridge
     void serialize_sensor_msgs_msg_laser_scan(const SensorMsgsMsgLaserScan &msg, std::vector<char> &buffer)
     {
         buffer.clear();
-        size_t offset = 0;
+        uint32_t offset = 0;
 
         // std_msgs/msg/Header frame_id size
-        size_t header_frame_id_size = msg.header_frame_id.size();
+        uint32_t header_frame_id_size = static_cast<uint32_t>(msg.header_frame_id.size());
         buffer.resize(offset + sizeof(header_frame_id_size));
         memcpy(buffer.data() + offset, &header_frame_id_size, sizeof(header_frame_id_size));
         offset += sizeof(header_frame_id_size);
@@ -69,7 +69,7 @@ namespace tcp_ip_bridge
         offset += sizeof(msg.range_max);
 
         // sensor_msgs/msg/LaserScan ranges size
-        size_t ranges_size = msg.ranges.size();
+        uint32_t ranges_size = static_cast<uint32_t>(msg.ranges.size());
         buffer.resize(offset + sizeof(ranges_size));
         memcpy(buffer.data() + offset, &ranges_size, sizeof(ranges_size));
         offset += sizeof(ranges_size);
@@ -80,7 +80,7 @@ namespace tcp_ip_bridge
         offset += ranges_size * sizeof(float);
 
         // sensor_msgs/msg/LaserScan intensities size
-        size_t intensities_size = msg.intensities.size();
+        uint32_t intensities_size = static_cast<uint32_t>(msg.intensities.size());
         buffer.resize(offset + sizeof(intensities_size));
         memcpy(buffer.data() + offset, &intensities_size, sizeof(intensities_size));
         offset += sizeof(intensities_size);
@@ -93,10 +93,10 @@ namespace tcp_ip_bridge
 
     void deserialize_sensor_msgs_msg_laser_scan(const std::vector<char> &buffer, SensorMsgsMsgLaserScan &msg)
     {
-        size_t offset = 0;
+        uint32_t offset = 0;
 
         // std_msgs/msg/Header frame_id size
-        size_t header_frame_id_size;
+        uint32_t header_frame_id_size;
         memcpy(&header_frame_id_size, buffer.data() + offset, sizeof(header_frame_id_size));
         offset += sizeof(header_frame_id_size);
 
@@ -141,7 +141,7 @@ namespace tcp_ip_bridge
         offset += sizeof(msg.range_max);
 
         // sensor_msgs/msg/LaserScan ranges size
-        size_t ranges_size;
+        uint32_t ranges_size;
         memcpy(&ranges_size, buffer.data() + offset, sizeof(ranges_size));
         offset += sizeof(ranges_size);
 
@@ -151,7 +151,7 @@ namespace tcp_ip_bridge
         offset += ranges_size * sizeof(float);
 
         // sensor_msgs/msg/LaserScan intensities size
-        size_t intensities_size;
+        uint32_t intensities_size;
         memcpy(&intensities_size, buffer.data() + offset, sizeof(intensities_size));
         offset += sizeof(intensities_size);
